// src/models/planetModel.ts
export type Planet = {
  name: string;
  rotation_period: number;
  orbital_period: string;
  diameter: string;
  climate: string;
  gravity: string;
  terrain: string;
  surface_water: string;
  population: string;
  residents: string[];
  films: string[];
  created: string;
  edited: string;
  url: string;
};

export type WookiePlanet = {
  whrascwo: string;
  rcooaoraaoahoowh_akworcahoowa: number;
  oorcrhahaoraan_akworcahoowa: string;
  waahrascwoaoworc: string;
  oaanahscraaowo: string;
  rrrcrahoahaoro: string;
  aoworcrcraahwh: string;
  churcwwraoawo_ohraaoworc: string;
  akooakhuanraaoahoowh: string;
  rcwocahwawowhaoc: string[];
  wwahanscc: string[];
  oarcworaaowowa: string;
  wowaahaowowa: string;
  hurcan: string;
};